requirements.txt
# Iron Doom Jarvis Dependencies

# Core Discord Bot
discord.py>=2.3.2
aiohttp>=3.8.5

# Scheduling
APScheduler>=3.10.4
pytz>=2023.3

# API Clients
notion-client>=2.2.1
requests>=2.31.0

# Data Processing & ML
pandas>=2.1.1
scikit-learn>=1.3.0
numpy>=1.24.3

# Environment & Configuration
python-dotenv>=1.0.0

# Logging & Utilities
colorlog>=6.7.0

# Optional AI Integration
openai>=0.28.1
google-genai>=0.7.2

# Development & Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1
black>=23.7.0
flake8>=6.0.0

# Production Server (Optional)
gunicorn>=21.2.0
uvicorn>=0.23.2